Remove Xr to system man pages
These create broken links when generating the manual page and there is
no way to selectively disable link generation in mandoc.

This change should not create much of an impact to readers on console,
who would enter a man command for the respective page regardless.

This also fixes a small usage error of the Xr macro in
crypto_aead_lock.3monocypher.