Created a dedicated dist/ directory for end user releases
Related to #14 and #24.  The repository is now unfit for distribution
to end users.  We must first execute the dist.sh script to generate
test vectors and move the relevant files to the dist/ directory, which
can then be archived and distributed.  This archive will contain a
makefile to compile and test Monocypher.

Stuff left to do:

- Review and finalise the tests.
- Generate html files for the manual.
- Revise the frama-c.sh and coverage.sh scripts
- Have the makefile install Monocypher
- Automate archive generation (and print the archive's blake2 hash).