Prefixed optional files with "monocypher-"
Monocypher is a "single file" library, with optional files. As a single
file library, it is best to occupy the global namespace directly. It's
only one include, no need for a folder.

The optional file kinda throw things off. We could put it in a
sub-folder, but we probably want both header to be at the same
place. And we certainly don't want to put monocypher.h itself in
sub-folder.

The solution is to have those files, none in a sub folder.

    monocypher.h
    monocypher.c
    monocypher-ed25519.h  (optional)
    monocypher-ed25519.c  (optional)

Namespace pollution is limited to a prefix.  Should be reasonable both
for projects that import those files directly, or for packaging in a
Linux or BSD distribution.